Enable running pytest examples, and refactor wgpu adapter checking
