+++ examples/tests/test_examples.py
@@ -13,16 +13,19 @@ import numpy as np
 import pytest
 
 
-from tests.testutils import (
+from testutils import (
     can_use_wgpu_lib,
-    wgpu_backend,
-    is_lavapipe,
+    adapter_summary,
     find_examples,
     ROOT,
     screenshots_dir,
     diffs_dir,
 )
 
+
+is_lavapipe = adapter_summary and all(
+    x in adapter_summary.lower() for x in ("llvmpipe", "vulkan")
+)
 
 if not can_use_wgpu_lib:
     pytest.skip("Skipping tests that need the wgpu lib", allow_module_level=True)
@@ -57,7 +60,7 @@ def mock_time():
 
 
 def test_that_we_are_on_lavapipe():
-    print(wgpu_backend)
+    print(adapter_summary)
     if os.getenv("EXPECT_LAVAPIPE"):
         assert is_lavapipe
