+++ tests/testutils.py
@@ -4,15 +4,9 @@ import sys
 import logging
 import subprocess
 from io import StringIO
-from pathlib import Path
 
+import wgpu
 from wgpu.utils import get_default_device  # noqa: F401 - imported by tests
-
-
-ROOT = Path(__file__).parent.parent  # repo root
-examples_dir = ROOT / "examples"
-screenshots_dir = examples_dir / "screenshots"
-diffs_dir = screenshots_dir / "diffs"
 
 
 class LogCaptureHandler(logging.StreamHandler):
@@ -66,21 +60,13 @@ def iters_equal(iter1, iter2):
     return False
 
 
-def _determine_can_use_wgpu_lib():
-    # For some reason, since wgpu-native 5c304b5ea1b933574edb52d5de2d49ea04a053db
-    # the process' exit code is not zero, so we test more pragmatically.
-    code = "import wgpu.utils; wgpu.utils.get_default_device(); print('ok')"
-    result = subprocess.run(
-        [sys.executable, "-c", code],
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        universal_newlines=True,
-    )
-    print("_determine_can_use_wgpu_lib() status code:", result.returncode)
-    return (
-        result.stdout.strip().endswith("ok")
-        and "traceback" not in result.stderr.lower()
-    )
+def get_default_adapter_summary():
+    """Get description of adapter, or None when no adapter is available."""
+    try:
+        adapter = wgpu.gpu.request_adapter_sync()
+    except RuntimeError:
+        return None  # lib not available, or no adapter on this system
+    return adapter.summary
 
 
 def _determine_can_use_glfw():
@@ -93,46 +79,9 @@ def _determine_can_use_glfw():
         return True
 
 
-def get_wgpu_backend():
-    """
-    Query the configured wgpu backend driver.
-    """
-    code = "import wgpu.utils; info = wgpu.utils.get_default_device().adapter.info; print(info['adapter_type'], info['backend_type'])"
-    result = subprocess.run(
-        [
-            sys.executable,
-            "-c",
-            code,
-        ],
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        universal_newlines=True,
-        cwd=ROOT,
-    )
-    out = result.stdout.strip()
-    err = result.stderr.strip()
-    return err if "traceback" in err.lower() else out
+adapter_summary = get_default_adapter_summary()
+can_use_wgpu_lib = bool(adapter_summary)
 
-
-def find_examples(query=None, negative_query=None, return_stems=False):
-    result = []
-    for example_path in examples_dir.glob("*.py"):
-        example_code = example_path.read_text()
-        query_match = query is None or query in example_code
-        negative_query_match = (
-            negative_query is None or negative_query not in example_code
-        )
-        if query_match and negative_query_match:
-            result.append(example_path)
-    result = list(sorted(result))
-    if return_stems:
-        result = [r.stem for r in result]
-    return result
-
-
-can_use_wgpu_lib = _determine_can_use_wgpu_lib()
 can_use_glfw = _determine_can_use_glfw()
 is_ci = bool(os.getenv("CI", None))
 is_pypy = sys.implementation.name == "pypy"
-wgpu_backend = get_wgpu_backend()
-is_lavapipe = wgpu_backend.lower() == "cpu vulkan"
